+++ pyxel/models/charge_generation/__init__.py
@@ -13,3 +13,4 @@ from .dark_current import dark_current_rule07
 from .load_profile import charge_profile
 from .photoelectrons import monte_carlo_conversion, simple_conversion, random_pos
 from .tars.tars import run_tars
+from .qe_map import qe_map