+++ CHANGELOG.rst
@@ -24,6 +24,8 @@ Core
   (See `#65 <https://gitlab.com/esa/pyxel/issues/65>`_).
 * Added a model for the amplifier crosstalk.
   (See `#116 <https://gitlab.com/esa/pyxel/issues/116>`_).
+* Added ability to load custom QE maps.
+  (See `#117 <https://gitlab.com/esa/pyxel/issues/117>`_).
 
 Others
 ~~~~~~